Add elm-http-extra
Just wanted to throw my own elm-http-extra in if you'll take it :smiley:

The design for the API came out of a discussion in the Slack channel, and we've got positive feedback from those that use it! I'm hoping more folks will adopt it and help improve it.

Update README.md