Problem: Potentially confusing reference to QML
Background:

PR #134 from @ShalokShalom introduced a link to QML with the idea that one could make mobile and desktop applications with Elm. QML is very cool, and I really want this to be true, but unfortunately there seems to be no real-world example (I searched!)

There was a similar PR to Fable by same author who admits that it is currently only a hypothetical possibility:

https://github.com/fable-compiler/fable-compiler.github.io/issues/14#issuecomment-335104505

In the absence of a concrete example of using QML with Elm this link to QML is just going to confuse and distract people.

I propose removing this link.